Rename package to meet style standards conventions
This patch renames the package to meet the style standards consisting of
lowercase characters without underscores. This will remove the warnings
produced by CheckStyle when following our custom style format.
