+++ app/src/androidTest/java/com/example/periodtracker/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.group_5_project;
+package com.example.periodtracker;
 
 import android.content.Context;
