+++ app/src/test/java/com/example/periodtracker/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.group_5_project;
+package com.example.periodtracker;
 
 import org.junit.Test;
